Add CF_RETURNS_NOT_RETAINED to DKDrawing's Export category methods that return CGImages. Quiet a couple more warnings.

+++ Source/DKDrawableShape+Hotspots.m
@@ -191,7 +191,7 @@
 
 - (void)endMouseTracking:(NSEvent*)event inView:(NSView*)view
 {
-	LogEvent_(kReactiveEvent, @"hotspot stopped tracking, partcode = %d", m_partcode);
+	LogEvent_(kReactiveEvent, @"hotspot stopped tracking, partcode = %ld", (long)m_partcode);
 
 	if ([self delegate] && [[self delegate] respondsToSelector:@selector(hotspot:
 																   didEndTrackingWithEvent:
